rhashtable: make test actually random
The "random rhlist add/delete operations" actually wasn't very random,
as all cases tested the same bit. Since the later parts of this loop
depend on the first case execute this unconditionally, and then test on
different bits for the remaining tests. While at it only request as much
random bits as are actually used.

Signed-off-by: Rolf Eike Beer <eike-kernel@sf-tec.de>
Acked-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>